+++ github-projects-md/scripts/config/categoryMap.js
@@ -1,7 +1,0 @@
-export const categoryMap = {
-  pro: "⚡️ Projets professionnels",
-  default: "🚀 Projets personnels",
-  SCPT: "⚙️ Scripting",
-  SB: "🧪 SandBox",
-  OC: "🎓 OpenClassrooms",
-};